refactor(tests): extract from_text_equivalence.rs (2 tests)
Move from_text_equivalence_* tests into engine/tests/from_text_equivalence.rs.
These are sync #[test] (not #[tokio::test]) equivalence tests between
DocumentIndex::from_text and fallback_scan_with_frontmatter.

Tests moved (now at engine::tests::from_text_equivalence::*):
  from_text_equivalence_with_fallback_scan_mixed_doc
  from_text_equivalence_frontmatter_only_doc

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

marky-n1h step 8 (3/5).
